+++ smartbrain-api/server.js
@@ -4,6 +4,7 @@ const bcrypt = require('bcrypt-nodejs');
 const cors = require('cors');
 const knex = require('knex');
 
+const signIn = require('./controllers/signin');
 const register = require('./controllers/register');
 
 const db = knex({
@@ -24,24 +25,7 @@ app.get('/', (req, res) => {
   res.send(database.users);
 })
 
-app.post('/signin', (req, res) => {
-  db.select('email', 'hash').from('login')
-    .where('email', '=', req.body.email)
-    .then(data => {
-      const isValid = bcrypt.compareSync(req.body.password, data[0].hash);
-      if (isValid) {
-        return db.select('*').from('users')
-          .where('email', '=', req.body.email)
-          .then(user => {
-            res.json(user[0])
-          })
-          .catch(err => res.status(400).json('Unable to get user'))
-      } else {
-        res.status(400).json('Wrong credentials')
-      }
-    })
-    .catch(err => res.status(400).json('Wrong credentials'))
-})
+app.post('/signin', (req, res) => { signIn.handleSignIn(req, res, db, bcrypt) })
 
 app.post('/register', (req, res) => { register.handleRegister(req, res, db, bcrypt) })
